requirements.txt
certifi==2016.2.28
configparser==3.5.0
enum34==1.1.6
feedparser==5.2.1
flake8==3.0.3
funcsigs==1.0.2
future==0.15.2
mccabe==0.5.2
mock==2.0.0
pbr==1.10.0
pluggy==0.3.1
py==1.4.31
pycodestyle==2.0.0
pyflakes==1.2.3
pymongo==3.3.0
pytest-mock==1.1
python-telegram-bot==5.0.0
requests==2.10.0
six==1.10.0
tinydb==3.2.1
tox==2.3.1
urllib3==1.16
virtualenv==15.0.2
wheel==0.24.0
